Add new buttons for day cash counts
